CharacterControllerの階層を変え、初期化用の分離クラスの追加
