+++ Assets/TechC/VBattle/Scripts/InGame/Player/CharacterConotoller/CharacterController_Buff.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f059ef5c5bf349a4ba9db87578925ee6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
