+++ Assets/TechC/VBattle/Scripts/InGame/Player/CharacterConotoller/CharacterController_Init.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a823ebd8ebc1a8246977db47a5345965
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
